+++ src/main/java/io/electrum/moneytransfer/model/MoneyTransferFeeQuote.java
@@ -25,6 +25,9 @@ public class MoneyTransferFeeQuote {
    @JsonProperty("receiver")
    private Institution receiver = null;
 
+   @JsonProperty("settlementEntity")
+   private Institution settlementEntity = null;
+
    @JsonProperty("feeQuote")
    private FeeQuote feeQuote = null;
 
@@ -82,6 +85,27 @@ public class MoneyTransferFeeQuote {
 
    public void setReceiver(Institution receiver) {
       this.receiver = receiver;
+   }
+
+   public MoneyTransferFeeQuote settlementEntity(Institution settlementEntity) {
+      this.settlementEntity = settlementEntity;
+      return this;
+   }
+
+   /**
+    * Data relating to the entity with whom the Merchant will settle the transaction..
+    *
+    * @return settlementEntity
+    **/
+   @JsonProperty("settlementEntity")
+   @ApiModelProperty(required = false, value = "Data relating to the entity with whom the Merchant will settle the transaction.")
+   @Valid
+   public Institution getSettlementEntity() {
+      return settlementEntity;
+   }
+
+   public void setSettlementEntity(Institution settlementEntity) {
+      this.settlementEntity = settlementEntity;
    }
 
    public MoneyTransferFeeQuote feeQuote(FeeQuote feeQuote) {